requirements.txt
python-dotenv
pandas
numpy
pymongo
certifi
pymongo[srv]>=4.0
scikit-learn
scipy
dill
pyaml
mlflow



# -e .